Adding lab03.ipynb format back
After second thoughts, @gallettilance thinks it is better to have it in a notebook than to conduct unit testing. He proposes in a future lecture, he may go over it instead. This will allow students not to be overwhelmed with the content required to be done within a 50-minute lab time span.